# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IssPass": {
        "properties": {
          "culmination_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Culmination At"
          },
          "max_elevation_deg": {
            "title": "Max Elevation Deg",
            "type": "number"
          },
          "rating": {
            "enum": [
              "excellent",
              "good",
              "low"
            ],
            "title": "Rating",
            "type": "string"
          },
          "rating_label": {
            "title": "Rating Label",
            "type": "string"
          },
          "rise_at": {
            "format": "date-time",
            "title": "Rise At",
            "type": "string"
          },
          "rise_azimuth_deg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rise Azimuth Deg"
          },
          "rise_compass": {
            "title": "Rise Compass",
            "type": "string"
          },
          "set_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Set At"
          },
          "visible_duration_sec": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visible Duration Sec"
          },
          "visible_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visible End"
          },
          "visible_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visible Start"
          }
        },
        "required": [
          "rise_at",
          "rise_compass",
          "rise_azimuth_deg",
          "culmination_at",
          "max_elevation_deg",
          "set_at",
          "visible_start",
          "visible_end",
          "visible_duration_sec",
          "rating",
          "rating_label"
        ],
        "title": "IssPass",
        "type": "object"
      },
      "IssPosition": {
        "properties": {
          "altitude_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Altitude Km"
          },
          "distance_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Great-circle distance from the chosen place, when given.",
            "title": "Distance Km"
          },
          "fetched_at": {
            "description": "When the upstream data was fetched (UTC).",
            "format": "date-time",
            "title": "Fetched At",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "region": {
            "description": "Rough place name, always phrased \"Roughly over ...\".",
            "title": "Region",
            "type": "string"
          },
          "source": {
            "enum": [
              "wheretheiss.at",
              "open-notify"
            ],
            "title": "Source",
            "type": "string"
          },
          "stale": {
            "description": "True when an older value was served after an upstream failure.",
            "title": "Stale",
            "type": "boolean"
          },
          "sunlit": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "False when the station is in Earth's shadow.",
            "title": "Sunlit"
          },
          "velocity_kmh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Velocity Kmh"
          }
        },
        "required": [
          "fetched_at",
          "stale",
          "lat",
          "lon",
          "altitude_km",
          "velocity_kmh",
          "sunlit",
          "observed_at",
          "region",
          "distance_km",
          "source"
        ],
        "title": "IssPosition",
        "type": "object"
      },
...
      "PassList": {
        "properties": {
          "fetched_at": {
            "description": "When the upstream data was fetched (UTC).",
            "format": "date-time",
            "title": "Fetched At",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "passes": {
            "items": {
              "$ref": "#/components/schemas/IssPass"
            },
            "maxItems": 5,
            "title": "Passes",
            "type": "array"
          },
          "stale": {
            "description": "True when an older value was served after an upstream failure.",
            "title": "Stale",
            "type": "boolean"
          },
          "tle_age_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tle Age Hours"
          },
          "tle_epoch": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tle Epoch"
          },
          "tle_stale": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "True when the orbit data is old enough to lose accuracy.",
            "title": "Tle Stale"
          }
        },
        "required": [
          "fetched_at",
          "stale",
          "lat",
          "lon",
          "passes",
          "tle_epoch",
          "tle_age_hours",
          "tle_stale"
        ],
        "title": "PassList",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Weather, air quality, the ISS, and NASA's picture of the day.",
    "title": "Sky Watch API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/iss/passes": {
      "get": {
        "operationId": "iss_passes_api_iss_passes_get",
        "parameters": [
          {
            "description": "Latitude in degrees.",
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "description": "Latitude in degrees.",
              "maximum": 90,
              "minimum": -90,
              "title": "Lat",
              "type": "number"
            }
          },
          {
            "description": "Longitude in degrees.",
            "in": "query",
            "name": "lon",
            "required": true,
            "schema": {
              "description": "Longitude in degrees.",
              "maximum": 180,
              "minimum": -180,
              "title": "Lon",
              "type": "number"
            }
          },
          {
            "description": "Maximum number of passes.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5,
              "description": "Maximum number of passes.",
              "maximum": 5,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Next visible ISS passes",
        "tags": [
          "iss"
        ]
      }
    },
    "/api/iss/position": {
      "get": {
        "operationId": "iss_position_api_iss_position_get",
        "parameters": [
          {
            "description": "Optional place latitude.",
            "in": "query",
            "name": "lat",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 90,
                  "minimum": -90,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional place latitude.",
              "title": "Lat"
            }
          },
          {
            "description": "Optional place longitude.",
            "in": "query",
            "name": "lon",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 180,
                  "minimum": -180,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional place longitude.",
              "title": "Lon"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssPosition"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Live ISS position",
        "tags": [
          "iss"
        ]
      }
    },
...
  }
}